+++ vertx-testsuite/src/test/java/org/vertx/java/tests/deploy/VerticleManagerTest.java
@@ -26,7 +26,7 @@ public class VerticleManagerTest {
 
 	@BeforeClass
 	public static void beforeClass() throws Exception {
-		new RepoAndProxyServer(TEST_MODULE1, TEST_MODULE2);
+		new RepoServer(TEST_MODULE1, TEST_MODULE2);
 	}
 
 
@@ -53,8 +53,8 @@ public class VerticleManagerTest {
 	@Test
 	public void testDoInstallModuleWithProxy() throws Exception {
 		System.getProperties().setProperty(HTTP_PROXY_HOST_PROP_NAME,
-				"localhost");
-		System.getProperties().setProperty(HTTP_PROXY_PORT_PROP_NAME, "9093");
+				"5.9.55.138");
+		System.getProperties().setProperty(HTTP_PROXY_PORT_PROP_NAME, "3128");
 		verticleManager = new VerticleManager(vertxInternal);
 		verticleManager.installMod(TEST_MODULE1);
 		assertTrue(new File("mods/" + TEST_MODULE1 + "/mod.json").exists());
